// src/ts/Interfaces.ts
export enum IPages
{
    home = '#page-home',
    dailies = '#page-dailies',
    projects = '#page-projects',
    about = '#page-about'
}

export enum ICategory
{
    experimental = 'experimental',
    interactive = 'interactive',
    Touchdesigner = 'Touchdesigner',
    DavinciResolve = 'DavinciResolve',
    Typescript = 'Typescript',
    Python = 'Python',
    Processing = 'Processing',
    Blender = 'Blender',
    Houdini = 'Houdini',
    SuperCollider = 'SuperCollider',
    Krita = 'Krita',
    Affinity_Designer = 'Affinity Designer',
    Affinity_Photo = 'Affinity Photo',
    Affinity_Publisher = 'Affinity Publisher'
}